Changes and Progress
In the gameMode class, i commented out some stuff that we have yet
to create, since its giving us errors. In Main Application, i created
the molePane to call the screen where the game will be played. I also
changed the molePane to be called first when running main app so i don't
have to click through the menus. In the Mole.java, i made the mainApp so
that it can be called, and added the game background just to get
everything started.